implement headline responsive props
